feature/WA-14 side menu code refactoring
