+++ src/components/SideMenu/SideMenu.module.css
@@ -1,12 +1,32 @@
 .sideMenu {
+  position: sticky;
+  top: 0;
   display: flex;
   flex-direction: column;
-  width: 80px;
-  height: 900px;
+  width: 150px;
+  height: 100vh;
   align-items: flex-start;
+  padding: 10px;
+  border-right: 1px solid rgba(255, 255, 255, 0.288);
+}
+
+.iconsLine {
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+}
+
+.cities {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  flex: 1;
+  color: white;
+  overflow: auto;
 }
 
 .city {
   font-size: xx-large;
   font-weight: lighter;
+  border-bottom: 1px solid rgba(255, 255, 255, 0.288);
 }
