+++ src/components/SideMenu/index.js
@@ -1,0 +1,1 @@
+export { default as SideMenu } from './SideMenu';
