+++ src/components/shared/IconButton/index.js
@@ -1,0 +1,1 @@
+export { default as IconButton } from './IconButton';
